Homework-4/microservices/technical_analysis/requirements.txt
flask
flask-cors
pandas
psycopg2-binary
ta
